% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistverse-package.R
\docType{package}
\name{redistverse-package}
\alias{redistverse}
\alias{redistverse-package}
\title{redistverse: Easily Install and Load Redistricting Software}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Easy installation, loading, and control of packages for redistricting data downloading, spatial data processing, simulation, analysis, and visualization. This package makes it easy to install and load multiple 'redistverse' packages at once. The 'redistverse' is developed and maintained by the Algorithm-Assisted Redistricting Methodology (ALARM) Project. For more details see \url{https://alarm-redist.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alarm-redist/redistverse}
  \item \url{https://alarm-redist.org/redistverse/}
  \item Report bugs at \url{https://github.com/alarm-redist/redistverse/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

Authors:
\itemize{
  \item Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})
}

Other contributors:
\itemize{
  \item tidyverse contributors (attach code adapted from tidyverse.) [contributor]
}

}
\keyword{internal}
