% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{redistverse_logo}
\alias{redistverse_logo}
\title{Print the 'redistverse' logo using ASCII characters}
\usage{
redistverse_logo()
}
\value{
\code{NULL}, invisibly
}
\description{
Print the 'redistverse' logo using ASCII characters
}
\examples{
redistverse_logo()
}
