% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{use_redistverse_r_universe}
\alias{use_redistverse_r_universe}
\title{Use \code{r-universe} Packages for \code{redistverse}}
\usage{
use_redistverse_r_universe()
}
\value{
formatted code to use \code{r-universe}
}
\description{
Use \code{r-universe} Packages for \code{redistverse}
}
\examples{
use_redistverse_r_universe()
}
