% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistverse-package.R
\docType{package}
\name{redistverse-package}
\alias{redistverse}
\alias{redistverse-package}
\title{redistverse: Easily Install and Load Redistricting Software}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'redistverse' is a set of packages for redistricting data downloading, spatial preprocessing, simulation, analysis, and visualization. This package makes it easy to install and load multiple 'redistverse' packages at once.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alarm-redist/redistverse}
  \item \url{http://alarm-redist/redistverse/}
  \item Report bugs at \url{https://github.com/alarm-redist/redistverse/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu}

Authors:
\itemize{
  \item Cory McCartan \email{cmccartan@g.harvard.edu}
}

Other contributors:
\itemize{
  \item tidyverse contributors (attach code adapted from tidyverse.) [contributor]
}

}
\keyword{internal}
